pix2pix/requirements.txt
torch
torchvision
pillow
matplotlib
numpy